+++ toolkit/scaffold/project/golang/__init__.py
@@ -1,0 +1,5 @@
+from .example import create_example_console
+
+__all__ = [
+    "create_example_console",
+]