+++ toolkit/template/scaffold/project/cpp/qt5/example/qml/main.js
@@ -1,7 +1,0 @@
-function init(window, core) {
-
-}
-
-function launch(window, core) {
-
-}