+++ toolkit/template/scaffold/project/cpp/qt5/example/qml/main.qml
@@ -3,9 +3,6 @@ import QtQuick.Window 2.12
 import QtQuick.Controls 2.12
 import QtQuick.Dialogs 1.1
 import QtQuick.Layouts 1.1
-import Qt.labs.platform 1.1
-
-import "main.js" as MainJS
 
 Window {
     id: window
@@ -15,7 +12,6 @@ Window {
     title: qsTr("{{ project_slug.words_capitalized }}")
 
     Component.onCompleted: {
-        MainJS.init(window, core)
-        MainJS.launch(window, core)
+
     }
 }